requirements.txt
requests
flask
flask_api
python-dotenv
numpy
pandas
scikit-surprise
psycopg2
redis